Adding PIT and EGA scrolling support
1. Added a code for setting up PIT and a small demonstration of the clock showing the working time.
2. Added scrolling of the EGA text buffer.

+++ kernel/arch/i386/inlineassembly/inlineassembly.c
@@ -40,7 +40,7 @@ inline uint32_t save_irqdisable(void)
     return flags;
 }
 
-inline void irqrestore(unsigned long flags)
+inline void irqrestore(uint32_t flags)
 {
     asm ("push %0\n\tpopf" : : "rm"(flags) : "memory","cc");
 }
